+++ src/lotus_json/mod.rs
@@ -123,527 +123,23 @@ where
 // but quickcheck is not powerful enough... we should use proptest instead/in addition
 
 use self::cid::CidLotusJson;
-mod cid {
-    use super::*;
-
-    pub type CidLotusJson = CidLotusJsonGeneric<64>;
-
-    #[derive(Serialize, Deserialize, From, Into)]
-    pub struct CidLotusJsonGeneric<const S: usize> {
-        #[serde(rename = "/", with = "stringify")]
-        slash: ::cid::CidGeneric<S>,
-    }
-
-    impl<const S: usize> HasLotusJson for ::cid::CidGeneric<S> {
-        type LotusJson = CidLotusJsonGeneric<S>;
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(json!({"/": "baeaaaaa"}), ::cid::Cid::default());
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: ::cid::Cid) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod cid;
 use token_amount::TokenAmountLotusJson;
-mod token_amount {
-    use super::*;
-
-    use crate::shim::econ::TokenAmount;
-
-    #[derive(Serialize, Deserialize)]
-    #[serde(transparent)]
-    pub struct TokenAmountLotusJson {
-        #[serde(with = "stringify")]
-        attos: num::BigInt,
-    }
-
-    impl HasLotusJson for TokenAmount {
-        type LotusJson = TokenAmountLotusJson;
-    }
-
-    impl From<TokenAmount> for TokenAmountLotusJson {
-        fn from(value: TokenAmount) -> Self {
-            Self {
-                attos: value.atto().clone(),
-            }
-        }
-    }
-
-    impl From<TokenAmountLotusJson> for TokenAmount {
-        fn from(value: TokenAmountLotusJson) -> Self {
-            Self::from_atto(value.attos)
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(json!("1"), TokenAmount::from_atto(1));
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: TokenAmount) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod token_amount;
 use address::AddressLotusJson;
-mod address {
-    use super::*;
-
-    use crate::shim::address::Address;
-
-    #[derive(Serialize, Deserialize, From, Into)]
-    #[serde(transparent)]
-    pub struct AddressLotusJson(#[serde(with = "stringify")] Address);
-
-    impl HasLotusJson for Address {
-        type LotusJson = AddressLotusJson;
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(json!("f00"), Address::default());
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: Address) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod address;
 use message::MessageLotusJson;
-mod message {
-    use super::{raw_bytes::RawBytesLotusJson, *};
-
-    use crate::shim::message::Message;
-
-    // TODO(aatifsyed): derive
-    #[derive(Serialize, Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    pub struct MessageLotusJson {
-        version: u64,
-        to: AddressLotusJson,
-        from: AddressLotusJson,
-        nonce: u64,
-        value: TokenAmountLotusJson,
-        gas_limit: u64,
-        gas_fee_cap: TokenAmountLotusJson,
-        gas_premium: TokenAmountLotusJson,
-        method: u64,
-        params: Option<RawBytesLotusJson>,
-        #[serde(rename = "CID", skip_serializing_if = "Option::is_none")]
-        cid: Option<CidLotusJson>,
-    }
-
-    impl HasLotusJson for Message {
-        type LotusJson = MessageLotusJson;
-    }
-
-    // TODO(aatifsyed): derive
-    impl From<MessageLotusJson> for Message {
-        fn from(value: MessageLotusJson) -> Self {
-            let MessageLotusJson {
-                version,
-                to,
-                from,
-                nonce,
-                value,
-                gas_limit,
-                gas_fee_cap,
-                gas_premium,
-                method,
-                params,
-                cid: _ignored, // TODO(aatifsyed): is this an error?
-            } = value;
-            Self {
-                version,
-                from: from.into(),
-                to: to.into(),
-                sequence: nonce,
-                value: value.into(),
-                method_num: method,
-                params: params.map(Into::into).unwrap_or_default(),
-                gas_limit,
-                gas_fee_cap: gas_fee_cap.into(),
-                gas_premium: gas_premium.into(),
-            }
-        }
-    }
-
-    impl From<Message> for MessageLotusJson {
-        fn from(value: Message) -> Self {
-            let Message {
-                version,
-                from,
-                to,
-                sequence,
-                value,
-                method_num,
-                params,
-                gas_limit,
-                gas_fee_cap,
-                gas_premium,
-            } = value;
-            Self {
-                version,
-                to: to.into(),
-                from: from.into(),
-                nonce: sequence,
-                value: value.into(),
-                gas_limit,
-                gas_fee_cap: gas_fee_cap.into(),
-                gas_premium: gas_premium.into(),
-                method: method_num,
-                params: Some(params.into()),
-                cid: None, // TODO(aatifsyed): is this an error?
-            }
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(
-            json!({
-                "From": "f00",
-                "GasFeeCap": "0",
-                "GasLimit": 0, // BUG?(aatifsyed)
-                "GasPremium": "0",
-                "Method": 0,
-                "Nonce": 0,
-                "Params": "", // BUG?(aatifsyed)
-                "To": "f00",
-                "Value": "0",
-                "Version": 0
-            }),
-            Message::default(),
-        );
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: Message) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod message;
 use signature_type::SignatureTypeLotusJson;
-mod signature_type {
-    use super::*;
-    use crate::shim::crypto::SignatureType;
-
-    #[derive(Deserialize, Serialize)]
-    #[serde(rename_all = "lowercase")]
-    pub enum SignatureTypeLotusJson {
-        Bls,
-        Secp256k1,
-        Delegated,
-    }
-
-    impl HasLotusJson for SignatureType {
-        type LotusJson = SignatureTypeLotusJson;
-    }
-
-    impl From<SignatureTypeLotusJson> for SignatureType {
-        fn from(value: SignatureTypeLotusJson) -> Self {
-            match value {
-                SignatureTypeLotusJson::Bls => Self::Bls,
-                SignatureTypeLotusJson::Secp256k1 => Self::Secp256k1,
-                SignatureTypeLotusJson::Delegated => Self::Delegated,
-            }
-        }
-    }
-
-    impl From<SignatureType> for SignatureTypeLotusJson {
-        fn from(value: SignatureType) -> Self {
-            match value {
-                SignatureType::Secp256k1 => Self::Secp256k1,
-                SignatureType::Bls => Self::Bls,
-                SignatureType::Delegated => Self::Delegated,
-            }
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(json!("bls"), SignatureType::Bls);
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: SignatureType) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod signature_type;
 use signature::SignatureLotusJson;
-mod signature {
-    use super::*;
-    use crate::shim::crypto::Signature;
-
-    #[derive(Serialize, Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    pub struct SignatureLotusJson {
-        r#type: SignatureTypeLotusJson,
-        #[serde(with = "base64_standard")]
-        data: Vec<u8>,
-    }
-
-    impl HasLotusJson for Signature {
-        type LotusJson = SignatureLotusJson;
-    }
-
-    impl From<SignatureLotusJson> for Signature {
-        fn from(value: SignatureLotusJson) -> Self {
-            let SignatureLotusJson { r#type, data } = value;
-            Self {
-                sig_type: r#type.into(),
-                bytes: data,
-            }
-        }
-    }
-
-    impl From<Signature> for SignatureLotusJson {
-        fn from(value: Signature) -> Self {
-            let Signature { sig_type, bytes } = value;
-            Self {
-                r#type: sig_type.into(),
-                data: bytes,
-            }
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(
-            json!({"Type": "bls", "Data": "aGVsbG8gd29ybGQh"}),
-            Signature {
-                sig_type: crate::shim::crypto::SignatureType::Bls,
-                bytes: Vec::from_iter(*b"hello world!"),
-            },
-        );
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: Signature) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
+mod signature;
 use signed_message::SignedMessageLotusJson;
-mod signed_message {
-    use crate::message::SignedMessage;
-
-    use super::*;
-
-    #[derive(Serialize, Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    pub struct SignedMessageLotusJson {
-        message: MessageLotusJson,
-        signature: SignatureLotusJson,
-        #[serde(rename = "CID", skip_serializing_if = "Option::is_none")]
-        cid: Option<CidLotusJson>,
-    }
-
-    impl HasLotusJson for SignedMessage {
-        type LotusJson = SignedMessageLotusJson;
-    }
-
-    impl From<SignedMessage> for SignedMessageLotusJson {
-        fn from(value: SignedMessage) -> Self {
-            let SignedMessage { message, signature } = value;
-            Self {
-                message: message.into(),
-                signature: signature.into(),
-                cid: None, // BUG?(aatifsyed)
-            }
-        }
-    }
-
-    impl From<SignedMessageLotusJson> for SignedMessage {
-        fn from(value: SignedMessageLotusJson) -> Self {
-            let SignedMessageLotusJson {
-                message,
-                signature,
-                cid: _ignored, // BUG?(aatifsyed)
-            } = value;
-            Self {
-                message: message.into(),
-                signature: signature.into(),
-            }
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(
-            json!({
-                "Message": {
-                    "From": "f00",
-                    "GasFeeCap": "0",
-                    "GasLimit": 0,
-                    "GasPremium": "0",
-                    "Method": 0,
-                    "Nonce": 0,
-                    "Params": "",
-                    "To": "f00",
-                    "Value": "0",
-                    "Version": 0
-                },
-                "Signature": {"Type": "bls", "Data": "aGVsbG8gd29ybGQh"}
-            }),
-            SignedMessage {
-                message: crate::shim::message::Message::default(),
-                signature: crate::shim::crypto::Signature {
-                    sig_type: crate::shim::crypto::SignatureType::Bls,
-                    bytes: Vec::from_iter(*b"hello world!"),
-                },
-            },
-        );
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: SignedMessage) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
-
-mod raw_bytes {
-    use super::*;
-    use fvm_ipld_encoding::RawBytes;
-
-    #[derive(Serialize, Deserialize)]
-    #[serde(transparent)]
-    pub struct RawBytesLotusJson(#[serde(with = "base64_standard")] Vec<u8>);
-
-    impl HasLotusJson for RawBytes {
-        type LotusJson = RawBytesLotusJson;
-    }
-
-    impl From<RawBytes> for RawBytesLotusJson {
-        fn from(value: RawBytes) -> Self {
-            RawBytesLotusJson(Vec::from(value))
-        }
-    }
-
-    impl From<RawBytesLotusJson> for RawBytes {
-        fn from(value: RawBytesLotusJson) -> Self {
-            Self::from(value.0)
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(
-            json!("aGVsbG8gd29ybGQh"),
-            RawBytes::new(Vec::from_iter(*b"hello world!")),
-        );
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: Vec<u8>) -> bool {
-            assert_via_json(RawBytes::new(val));
-            true
-        }
-    }
-}
-
-mod vec {
-    use super::*;
-
-    pub struct VecLotusJson<T>(Vec<T>);
-
-    impl<T> HasLotusJson for Vec<T>
-    where
-        T: HasLotusJson,
-    {
-        type LotusJson = VecLotusJson<T::LotusJson>;
-    }
-
-    impl<T> Serialize for VecLotusJson<T> {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            match self.0.is_empty() {
-                true => serializer.serialize_none(),
-                false => self.serialize(serializer),
-            }
-        }
-    }
-
-    impl<'de, T> Deserialize<'de> for VecLotusJson<T>
-    where
-        T: Deserialize<'de>,
-    {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            Option::<Vec<T>>::deserialize(deserializer)
-                .map(Option::unwrap_or_default)
-                .map(Self)
-        }
-    }
-
-    // VecLotusJson<T::LotusJson> -> Vec<T>
-    impl<T> From<VecLotusJson<T::LotusJson>> for Vec<T>
-    where
-        T: HasLotusJson,
-        T::LotusJson: Into<T>,
-    {
-        fn from(value: VecLotusJson<T::LotusJson>) -> Self {
-            value.0.into_iter().map(Into::into).collect()
-        }
-    }
-
-    // Vec<T> -> VecLotusJson<T::LotusJson>
-    impl<T> From<Vec<T>> for VecLotusJson<T::LotusJson>
-    where
-        T: HasLotusJson + Into<T::LotusJson>,
-    {
-        fn from(value: Vec<T>) -> Self {
-            Self(value.into_iter().map(Into::into).collect())
-        }
-    }
-
-    #[test]
-    fn test() {
-        assert_snapshot(json!([{"/": "baeaaaaa"}]), vec![::cid::Cid::default()]);
-    }
-
-    #[cfg(test)]
-    quickcheck! {
-        fn round_trip(val: Vec<::cid::Cid>) -> bool {
-            assert_via_json(val);
-            true
-        }
-    }
-}
+mod signed_message;
+use raw_bytes::RawBytesLotusJson;
+mod raw_bytes;
+use vec::VecLotusJson;
+mod vec;
 
 /// Usage: `#[serde(with = "stringify")]`
 mod stringify {
@@ -669,6 +165,7 @@ mod stringify {
     }
 }
 
+/// Usage: `#[serde(with = "base64_standard")]`
 mod base64_standard {
     use super::*;
